test(datetime): migrate highlighted-dates
